Refactor db tests
Refactor db tests to combine into a single file and make clear they test db CRUD only
